refactor: rename jest-mock to spy, allow new on spy, rename Context
