+++ packages/vitest/src/integrations/spy.ts
@@ -16,11 +16,12 @@ interface MockResultThrow {
 
 type MockResult<T> = MockResultReturn<T> | MockResultThrow | MockResultIncomplete
 
-export interface JestMockCompatContext<TArgs, TReturns> {
+export interface SpyContext<TArgs, TReturns> {
   calls: TArgs[]
   instances: TReturns[]
   invocationCallOrder: number[]
   results: MockResult<TReturns>[]
+  lastCall: TArgs | undefined
 }
 
 type Procedure = (...args: any[]) => any
@@ -38,7 +39,7 @@ type Classes<T> = {
 export interface SpyInstance<TArgs extends any[] = any[], TReturns = any> {
   getMockName(): string
   mockName(n: string): this
-  mock: JestMockCompatContext<TArgs, TReturns>
+  mock: SpyContext<TArgs, TReturns>
   mockClear(): this
   mockReset(): this
   mockRestore(): void
@@ -56,6 +57,7 @@ export interface SpyInstance<TArgs extends any[] = any[], TReturns = any> {
 
 export interface SpyInstanceFn<TArgs extends any[] = any, TReturns = any> extends SpyInstance<TArgs, TReturns> {
   (...args: TArgs): TReturns
+  new (...args: TArgs): TReturns
 }
 
 export type MaybeMockedConstructor<T> = T extends new (
@@ -163,6 +165,9 @@ function enhanceSpy<TArgs extends any[], TReturns>(
         const type = callType === 'error' ? 'throw' : 'return'
         return { type, value }
       })
+    },
+    get lastCall() {
+      return stub.calls.at(-1)
     },
   }
